client fix
